Extract tasks for reload, add systemctl
Even though it's only used once, I thought it would be nice to have an
encapsulated reload command.

Also, adding systemctl reload to this reload step. If you use systemd to
start monit, changes to config files are not recognized until you
restart the service.

This is particular to users of monit that manage the main monit process
via systnemd, so this change is probably not appropriate for the real
capistrano-sidekiq gem, but it is appropriate for our fork of that gem.